Simplify add command parser
