+++ src/main/java/seedu/address/logic/parser/AddCommandParser.java
@@ -7,22 +7,64 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
 
+import java.util.Set;
 import java.util.stream.Stream;
 
 import seedu.address.logic.commands.AddCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.model.ContactType;
+import seedu.address.model.client.Client;
+import seedu.address.model.contact.Address;
+import seedu.address.model.contact.Contact;
+import seedu.address.model.contact.Email;
+import seedu.address.model.contact.Name;
+import seedu.address.model.contact.Phone;
+import seedu.address.model.serviceprovider.ServiceProvider;
+import seedu.address.model.tag.Tag;
 
 /**
  * Parses input arguments and creates a new AddCommand object
  */
-public abstract class AddCommandParser implements Parser<AddCommand> {
+public class AddCommandParser implements Parser<AddCommand> {
+
+    private final ContactType contactType;
+
+    public AddCommandParser(ContactType contactType) {
+        this.contactType = contactType;
+    }
 
     /**
      * Parses the given {@code String} of arguments in the context of the AddCommand
      * and returns an AddCommand object for execution.
      * @throws ParseException if the user input does not conform the expected format
      */
-    public abstract AddCommand parse(String args) throws ParseException;
+    public AddCommand parse(String args) throws ParseException {
+        ArgumentMultimap argMultimap = createLegalArgumentMultimap(args);
+
+        Name name = ParserUtil.parseName(argMultimap.getValue(PREFIX_NAME).get());
+        Phone phone = ParserUtil.parsePhone(argMultimap.getValue(PREFIX_PHONE).get());
+        Email email = ParserUtil.parseEmail(argMultimap.getValue(PREFIX_EMAIL).get());
+        Address address = ParserUtil.parseAddress(argMultimap.getValue(PREFIX_ADDRESS).get());
+        Set<Tag> tagList = ParserUtil.parseTags(argMultimap.getAllValues(PREFIX_TAG));
+
+        Contact contact = new Client(name, phone, email, address, tagList);
+
+        return new AddCommand(contact);
+    }
+
+    /**
+     * Creates either a client or a service provider contact given the name, phone, email, address and tags.
+     */
+    private Contact createContact(Name name, Phone phone, Email email, Address address, Set<Tag> tagList) {
+        switch (contactType) {
+        case CLIENT:
+            return new Client(name, phone, email, address, tagList);
+        case SERVICE_PROVIDER:
+            return new ServiceProvider(name, phone, email, address, tagList);
+        default:
+            throw new RuntimeException("No such Contact Type");
+        }
+    }
 
     /**
      * Creates a {@code ArgumentMultimap} using the arguments from the input.
