+++ src/test/data/Account/accountlist.xml
@@ -12,7 +12,7 @@
     </accounts>
     <accounts>
         <username>whiterose</username>
-        <password>$31$16$EzsQPooNo8j_RoQ1NDLNAMbIQPol0IrdVRHcvWIAa24</password>
+        <password>$31$16$zYnrVv-TQNCR0sMQlbP8igbrDF0AGS0Qnc1j9dm2nmM</password>
         <role>READ_ONLY_USER</role>
     </accounts>
 </accountmanager>
