+++ src/test/data/XmlAccountStorageTest/accountlist.xml
@@ -2,17 +2,17 @@
 <accountmanager>
     <accounts>
         <username>rootUser</username>
-        <password>$31$16$54Xb94cY8KGmCuAoXb73DxuwHvAiGCM0N3X1s7f59sY</password>
+        <password>$31$16$9iIXrEtsP09SiJjAPTJBQ3c4QJr5nKZfVGWLv53e7Qs</password>
         <role>SUPER_USER</role>
     </accounts>
     <accounts>
         <username>simun</username>
-        <password>$31$16$oi3g0vixWTN1tAIgxTmwWi4aI_ChyBE5xDhovvU7V6k</password>
+        <password>$31$16$mzaH7NrtWwPFV3nRbsn2c95HNNIL4A483-iNmpYa4u0</password>
         <role>SUPER_USER</role>
     </accounts>
     <accounts>
         <username>whiterose</username>
-        <password>$31$16$zpfvx4WMCitRkyqDRHdcks_0ifXo1dUQ8HhApGsDWaQ</password>
+        <password>$31$16$3PXeyJn0VlXN8R_zHQzHcc62gXzuvtghdNFJY35drSg</password>
         <role>READ_ONLY_USER</role>
     </accounts>
 </accountmanager>
