+++ src/test/data/XmlAccountStorageTest/accountlistToChangePassword.xml
@@ -12,7 +12,7 @@
     </accounts>
     <accounts>
         <username>whiterose</username>
-        <password>$31$16$oSJKNWsb77a-xvUeaBNX1WgS7dHGBJLVRbgW1N-lKlA</password>
+        <password>$31$16$T3LSCx0J4psuwm0nSqLxdyXhGgg_b6pHiL_tTKb98FI</password>
         <role>READ_ONLY_USER</role>
     </accounts>
 </accountmanager>
