+++ src/test/data/XmlUtilTest/accountlist.xml
@@ -12,7 +12,7 @@
     </accounts>
     <accounts>
         <username>whiterose</username>
-        <password>$31$16$UDgJjc-x6B_OzgLTcT936negMRaSYgpWTYOT6cC7BiQ</password>
+        <password>$31$16$3LhXxsjl3uXMcJKI2hIBQa-AxEnBbnxPCDJ5khPRSIU</password>
         <role>READ_ONLY_USER</role>
     </accounts>
 </accountmanager>
